refactor(nextjs): move current setup to `src/pages` for convenience
