+++ packages/nextjs-router/src/pages/check-authentication.ts
